Frames, tracing, and dict comprehensions
- Added dict comprehensions
- Added stack o’ frames - each invocation of a func, list
comprehension, or dict comprehension gets its own set of locals.
There’s no support for the global statement yet, so assigning to
globals in those three scopes is not currently possible.
- Couple of new test scripts to make sure variables are/aren’t leaking
out of scopes as appropriate.
- Added tracing function and calls to support debugging interface.
- import statement (and derivatives) are accepted by the parser and
evaluator, but are NOOPs. This was done to allow for imports to appear
in the eval script to make PyCharm happy (not show everything as
undefined when defined in asteval globals).